requirements.txt
psycopg2-binary
python-dotenv
openmeteo-requests
requests-cache
retry-requests
pandas
numpy